requirements.txt
Django~=2.2.4
djangorestframework
db-sqlite3
